Timer: CA, Notifs: LeftSwipe, Clock: Down Swipe.
The plan is to have the task app in the center. This is to minimize distractions and 'stopped clock illusion.' In preparation a temporary app (timer) was put in the center where clock was. The clock was moved to the top and the notification was moved to the left. Inspired by Mac's Notifcation center. Also follows a convention of opening notifications and clearing them. If a problem arise the notification and clock can be switched.