i have a basic firmware with the start of new apps
right now I have the start of dice and the individual timer app that open(so not the main singlelane app[which has the colored buttons]). Reading through DisplayApp.cpp each screen handles it's own touch events so I'm going to have to copy the closest app. Probably alarm(-most functionality).Then redirect it to access data and setup the Timer based on that data.