Omg I never added any of the new files
technically speaking all of the new changes are not seen here seperatey but the summation of changes to these files is reflectd in this file. Also I saved because I changed from using a Application List base to a Tile base. The Tile is just one page of the App list and thats all I need. Also it gives me the control I want.